gui: panic handling p2
